requirements.txt
boto3==1.34.72
botocore==1.34.72
python-dotenv==1.0.1
